requirements.txt
pandas
pillow
PyQt5
